# specs/002-better-auth/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Better Auth API for AI Todo Chatbot (CORRECTED)
  version: 1.0.0
  description: API specification for backend services with Better Auth session verification

paths:
  /api/{user_id}/chat:
    post:
      summary: Chat with AI assistant (requires authenticated session)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user ID (must match authenticated user)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                conversation_id:
                  type: integer
                  nullable: true
                message:
                  type: string
              required:
                - message
      responses:
        '200':
          description: Chat response successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: integer
                  response:
                    type: string
                  tool_calls:
                    type: array
                    items:
                      type: object
        '401':
          description: Invalid or expired session
        '403':
          description: User ID in path doesn't match authenticated user
        '500':
          description: Server error

  /api/{user_id}/tasks:
    get:
      summary: Get user's tasks (requires authenticated session)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user ID (must match authenticated user)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        title:
                          type: string
                        completed:
                          type: boolean
                        created_at:
                          type: string
                          format: date-time
        '401':
          description: Invalid or expired session
        '403':
          description: User ID in path doesn't match authenticated user
        '500':
          description: Server error
    post:
      summary: Create a new task (requires authenticated session)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user ID (must match authenticated user)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                  nullable: true
              required:
                - title
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  description:
                    type: string
                  completed:
                    type: boolean
                  created_at:
                    type: string
                    format: date-time
        '401':
          description: Invalid or expired session
        '403':
          description: User ID in path doesn't match authenticated user
        '500':
          description: Server error

  /api/{user_id}/tasks/{task_id}:
    get:
      summary: Get a specific task (requires authenticated session)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user ID (must match authenticated user)
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The task ID
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  description:
                    type: string
                  completed:
                    type: boolean
                  created_at:
                    type: string
                    format: date-time
        '401':
          description: Invalid or expired session
        '403':
          description: User ID in path doesn't match authenticated user
        '404':
          description: Task not found
        '500':
          description: Server error
    put:
      summary: Update a task (requires authenticated session)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user ID (must match authenticated user)
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The task ID
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  nullable: true
                description:
                  type: string
                  nullable: true
                completed:
                  type: boolean
                  nullable: true
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  description:
                    type: string
                  completed:
                    type: boolean
                  created_at:
                    type: string
                    format: date-time
        '401':
          description: Invalid or expired session
        '403':
          description: User ID in path doesn't match authenticated user
        '404':
          description: Task not found
        '500':
          description: Server error
    delete:
      summary: Delete a task (requires authenticated session)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user ID (must match authenticated user)
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The task ID
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Invalid or expired session
        '403':
          description: User ID in path doesn't match authenticated user
        '404':
          description: Task not found
        '500':
          description: Server error

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Better Auth session token forwarded from frontend

tags:
  - name: chat
    description: AI chat functionality
  - name: tasks
    description: Task management functionality